RENAME: New response encoder/decoder
